llm temp

+++ StreamSchedule/LLM/Loader.cs
@@ -4,7 +4,7 @@ public static class Loader
 {
     private static readonly string directoryPath = Path.Combine(AppContext.BaseDirectory, "save");
     private const string ModelFileName = "model.bin";
-    
+
     public static (bool succ, int d, int l, int v, long count) LoadWeights()
     {
         string modelPath = Path.Combine(directoryPath, ModelFileName);
@@ -13,16 +13,16 @@ public static class Loader
             Console.WriteLine("no save for weights");
             return (false, 0, 0, 0, 0);
         }
-        
+
         long paramCount = 0;
-        
+
         using var modelReader = new BinaryReader(File.OpenRead(modelPath));
-        
+
         int d = modelReader.ReadInt32();
         int l = modelReader.ReadInt32();
         int v = modelReader.ReadInt32();
         Model.Initialize(d, l, v);
-        
+
         paramCount += ReadArray(modelReader, Model.OutputProjection);
         paramCount += ReadArray(modelReader, Model.OutputBiases);
         paramCount += ReadArray(modelReader, Model.Embedding);
@@ -45,6 +45,7 @@ public static class Loader
             paramCount += ReadArray(modelReader, Model.CM_WMixK[i]);
             paramCount += ReadArray(modelReader, Model.CM_WMixR[i]);
         }
+
         Model.CalculateDecay();
         Model.ParamCount = paramCount;
         return (true, d, l, v, paramCount);
@@ -52,7 +53,7 @@ public static class Loader
 
     private static long ReadArray(BinaryReader reader, float[] destination)
     {
-        for (long i = 0; i < destination.LongLength; i++) { destination[i] = reader.ReadSingle(); }
+        for (long i = 0; i < destination.LongLength; i++) { destination[i] = (float)reader.ReadHalf(); }
         return destination.LongLength;
     }
 }
